moved search related types to types/search
