+++ src/services/searchService.ts
@@ -1,21 +1,21 @@
 import * as chrono from 'chrono-node';
 import { getEndpoint } from 'utils/common/apiUtil';
 import { getToken } from 'utils/common/key';
-import { DateValue, Suggestion, SuggestionType } from 'components/SearchBar';
 import HTTPService from './HTTPService';
 import { Collection } from 'types/collection';
 import { EnteFile } from 'types/file';
 
 import { logError } from 'utils/sentry';
+import {
+    DateValue,
+    LocationSearchResponse,
+    Suggestion,
+    SuggestionType,
+} from 'types/search';
 
 const ENDPOINT = getEndpoint();
+
 const DIGITS = new Set(['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']);
-export type Bbox = [number, number, number, number];
-export interface LocationSearchResponse {
-    place: string;
-    bbox: Bbox;
-}
-export const getMapboxToken = () => process.env.NEXT_PUBLIC_MAPBOX_TOKEN;
 
 export function parseHumanDate(humanDate: string): DateValue[] {
     const date = chrono.parseDate(humanDate);
