+++ src/utils/search/index.ts
@@ -1,6 +1,5 @@
-import { DateValue } from 'components/SearchBar';
 import { EnteFile } from 'types/file';
-import { Bbox } from 'services/searchService';
+import { Bbox, DateValue } from 'types/search';
 
 export function isInsideBox(
     file: { longitude: number; latitude: number },
